+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Breaking**: `states` props is now just `state`.
 - State can now be undefined, it will not throw if not passed, the default value is `[{}, {}]`. It will still throw however if you passe an empty array `state={[]}` use `state={undefined}` instead if you want to do this.
 - State can now be static `state=[{ /* config */ }]` by passing a single state.
+- You can now pass an animation and layout preset from the available presets with the `preset`, `layoutPreset` prop.
 
 - You can now style the behavior view directly from the props (you still have the option to use the `style` prop too), any additional props are passed to the style object of the view, additional props are style overrides.